wb_price_parser/requirements.txt
# WB Price Parser - Dependencies
# Python 3.10+ required

# Environment variables
python-dotenv==1.0.0

# Async HTTP client
aiohttp==3.9.1

# HTML parsing
lxml==5.1.0

# Excel export
openpyxl==3.1.2
pandas==2.1.4

# Utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3

